pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "wolfkrypt-host"
version = "1.0.0"
description = "Wolfkrypt Screen Mirror - Windows/Linux/macOS host application"
authors = ["kevil123k"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
PyQt6 = "^6.8.0"
pyusb = "^1.3.1"
pynacl = "^1.5.0"
av = "^14.0.0"
sounddevice = "^0.5.1"
numpy = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-qt = "^4.4.0"
black = "^25.12.0"
flake8 = "^7.1.0"
isort = "^5.13.0"
pyinstaller = "^6.17.0"

[tool.poetry.scripts]
wolfkrypt-host = "src.main:main"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
